feat: added permissions for group of users

+++ controllers/bootcamps.js
@@ -54,15 +54,25 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route PUT /api/v1/bootcamps/:id
 // @access Private
 exports.editBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(
-        req.params.id,
-        req.body,
-        { new: true, runValidators: true }
-    );
+    let bootcamp = await Bootcamp.findById(req.params.id);
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp with id: ${req.params.id} is not found`, 404));
     }
+
+    // Check bootcamp ownership
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorize to edit this bootcamp`, 401)
+        );
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate( 
+        req.params.id,
+        req.body,
+        { new: true, runValidators: true }
+    )
+
     res.status(200).json({
         success: true,
         data: bootcamp
@@ -77,6 +87,13 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp with id of ${req.params.id} not found`, 404));
+    }
+
+    // Check bootcamp ownership
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorize to delete this bootcamp`, 401)
+        );
     }
 
     await bootcamp.deleteOne();
@@ -124,6 +141,13 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp with id of ${req.params.id} not found`, 404));
+    }
+
+    // Check bootcamp ownership
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorize to add photo to this bootcamp`, 401)
+        );
     }
 
     if (!req.files) {
