+++ controllers/courses.js
@@ -47,11 +47,21 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId)
 
     if (!bootcamp) {
         return next(new ErrorResponse(`No bootcamp with such id: ${req.params.bootcampId}`), 400);
+    }
+    // Check bootcamp ownership
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorize to add a course to bootcamp ${bootcamp._id}`
+                , 401
+            )
+        );
     }
 
     const course = await Course.create(req.body);
@@ -75,6 +85,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
             new ErrorResponse(`No course with such id: ${req.params.id}`),
             400
         );
+				}
+	
+				// Check course ownership
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorize to edit a course ${course._id}`
+                , 401
+            )
+        );
     }
 
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -97,6 +117,16 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
 
     if (!course) {
         return next(new ErrorResponse(`No course with such id: ${req.params.id}`), 400);
+				}
+	
+				// Check course ownership
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorize to delete a course ${course._id}`
+                , 401
+            )
+        );
     }
 
     await course.deleteOne();
